Fixed issue when used with grunt-contrib-watch
tihs.async()'s done function was never being called, and it was keeping
watched hjson files from being updated more than once (locking up watch
in general)

Per the grunt docs ( http://gruntjs.com/api/inside-tasks#this.async )
and the docs for async  (https://github.com/caolan/async ) the callback
for forEach has to be called on error and success. Modified the callback
to return the error to the done function (eg, next) and to call it with
success.